Remove `.clean()` from resize() with factor arg
The tests were failing with this error:

GD Warning: one parameter to a memory allocation multiplication is negative
or zero, failing operation gracefully

This is because the width and height values were effectively arbitrary and
it looked like the pointer was being used which was pointing to freed
memory.  Upon closer inspection, I noticed that the `resize` multi method
which is pointed to from the `resize` sub passes control to a more general
`resize` multimethod and within this method the image object is being
destroyed.  Even though the `.clean` method checks for the existence of an
image object before trying to destroy it again, it seems that it was calling
gdImageDestroy() an extra time, thus causing this problem.  Nevertheless,
removing the trailing `.clean()` makes the test suite run to completion and
the tests all pass.